experimental mdoc example dynamic extractor
this is a WIP for a mdoc example extractor which dynamically reads the
attributes from the MDOC using the witness code in longfellow-zk. It may or may
not work, with the added value of testing more the mDOC witness extractor. Left
here for later, right now the simple one with hard-coded attributes works well.